Show Address on the map instad of the case id. Added logging on catch when the json can not be parsed. Added emojis to show on the map

+++ HTB/v2/intranetx/routeplanner/dto/OpenRouteService.cs
@@ -115,7 +115,7 @@ namespace AddressAndDistanceLookup
         public int PlaceId { get; set; }
         public string licence { get; set; }
         public string osm_type { get; set; }
-        public int osm_id { get; set; }
+        public long osm_id { get; set; }
         public string @class { get; set; }
         public string type { get; set; }
         public int place_rank { get; set; }
